+++ crates/plugin_bundle/src/resource_pot_to_bundle/targets/generate.rs
@@ -211,7 +211,8 @@ pub fn generate_export_by_reference_export(
   is_already_polyfilled: &mut bool,
   options: &ShareBundleContext,
 ) -> Result<Vec<ModuleItem>> {
-  let mut patch_export_to_module: Vec<ModuleItem> = vec![];
+  let mut patch_export_to_module = vec![];
+
   if let Some(export) = bundle_reference.export.as_ref() {
     patch_export_to_module.extend(generate_export_as_module_export(
       should_reexport_raw,