+++ examples/vite-adapter-vue/src/pages/about.vue
@@ -8,6 +8,8 @@
 
 <script lang="tsx" setup>
 import { ElAlert } from 'element-plus';
+import 'element-plus/theme-chalk/el-alert.css'
+import 'element-plus/theme-chalk/el-button.css'
 
 const alert = <ElAlert style="margin-top: 30px;" title="About Page" type="success" description="This is jsx component." show-icon />;
 </script>